feat: prepare TestParser and TestWeeder for new parser
